requirements.txt
discord.py
opengradient
python-dotenv
